Convert timestamps to `Fluent::EventTime` to prevent type errors
What used to happen is that RabbitMQInput simply emitted the value of
properties['timestamp'], whose type is `Time` not `EventTime`. Since
input plugins are supposed to emit timestamps in `EventTime`, this was
causing unexpected exceptions (type errors) in the down stream.

This patch fixes the issue by coercing the type of timestamps properly
before emitting events.